+++ crates/node-core/Cargo.toml
@@ -15,7 +15,7 @@ workspace = true
 reth-primitives.workspace = true
 reth-fs-util.workspace = true
 reth-db = { workspace = true, features = ["mdbx"] }
-reth-storage-errors = { workspace = true, features = ["clap"] }
+reth-storage-errors.workspace = true
 reth-provider.workspace = true
 reth-network = { workspace = true, features = ["serde"] }
 reth-network-p2p.workspace = true
