+++ crates/storage/errors/Cargo.toml
@@ -15,7 +15,5 @@ reth-primitives.workspace = true
 reth-fs-util.workspace = true
 
 thiserror.workspace = true
-clap = { workspace = true, features = ["derive"], optional = true }
 
-[features]
-clap = ["dep:clap"]
+
